pam/integration-tests: Move all the generic helpers to the helpers file
These are not related to the cli tests anymore
