+++ pam/integration-tests/cli_test.go
@@ -1,19 +1,15 @@
 package main_test
 
 import (
-	"context"
-	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/msteinert/pam/v2"
 	"github.com/stretchr/testify/require"
 	"github.com/ubuntu/authd/internal/testutils"
 	localgroupstestutils "github.com/ubuntu/authd/internal/users/localgroups/testutils"
-	"github.com/ubuntu/authd/pam/internal/pam_test"
 )
 
 var daemonPath string
@@ -179,136 +175,6 @@ func TestPamCLIRunStandalone(t *testing.T) {
 	require.Contains(t, outStr, pam.ErrIgnore.Error())
 }
 
-func runAuthd(t *testing.T, gpasswdOutput, groupsFile string, currentUserAsRoot bool) string {
-	t.Helper()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	env := localgroupstestutils.AuthdIntegrationTestsEnvWithGpasswdMock(t, gpasswdOutput, groupsFile)
-	if currentUserAsRoot {
-		env = append(env, authdCurrentUserRootEnvVariableContent)
-	}
-	socketPath, stopped := testutils.RunDaemon(ctx, t, daemonPath, testutils.WithEnvironment(env...))
-	t.Cleanup(func() {
-		cancel()
-		<-stopped
-	})
-	return socketPath
-}
-
-func prepareClientTest(t *testing.T, clientPath string) []string {
-	t.Helper()
-
-	// Due to external dependencies such as `vhs`, we can't run the tests in some environments (like LP builders), as we
-	// can't install the dependencies there. So we need to be able to skip these tests on-demand.
-	if os.Getenv("AUTHD_SKIP_EXTERNAL_DEPENDENT_TESTS") != "" {
-		t.Skip("Skipping tests with external dependencies as requested")
-	}
-
-	pamCleanup, err := buildPAMTestClient(clientPath)
-	require.NoError(t, err, "Setup: Failed to build PAM executable")
-	t.Cleanup(pamCleanup)
-
-	return []string{
-		fmt.Sprintf("%s=%s", pam_test.ClientEnvExecModule, buildExecModule(t)),
-		fmt.Sprintf("%s=%s", pam_test.ClientEnvPath, buildPAMClient(t)),
-	}
-}
-
-// buildPAMTestClient builds the PAM module in a temporary directory and returns a cleanup function.
-func buildPAMTestClient(execPath string) (cleanup func(), err error) {
-	cmd := exec.Command("go", "build")
-	if testutils.CoverDirForTests() != "" {
-		// -cover is a "positional flag", so it needs to come right after the "build" command.
-		cmd.Args = append(cmd.Args, "-cover")
-	}
-	if testutils.IsAsan() {
-		// -asan is a "positional flag", so it needs to come right after the "build" command.
-		cmd.Args = append(cmd.Args, "-asan")
-	}
-	if testutils.IsRace() {
-		cmd.Args = append(cmd.Args, "-race")
-	}
-	cmd.Args = append(cmd.Args, "-tags=pam_binary_cli", "-o", filepath.Join(execPath, "pam_authd"), "../.")
-	if out, err := cmd.CombinedOutput(); err != nil {
-		return func() {}, fmt.Errorf("%v: %s", err, out)
-	}
-
-	return func() { _ = os.Remove(filepath.Join(execPath, "pam_authd")) }, nil
-}
-
-func buildPAMClient(t *testing.T) string {
-	t.Helper()
-
-	cmd := exec.Command("go", "build", "-C", "pam")
-	cmd.Dir = testutils.ProjectRoot()
-	if testutils.CoverDirForTests() != "" {
-		// -cover is a "positional flag", so it needs to come right after the "build" command.
-		cmd.Args = append(cmd.Args, "-cover")
-	}
-	if testutils.IsAsan() {
-		// -asan is a "positional flag", so it needs to come right after the "build" command.
-		cmd.Args = append(cmd.Args, "-asan")
-	}
-	if testutils.IsRace() {
-		cmd.Args = append(cmd.Args, "-race")
-	}
-	cmd.Args = append(cmd.Args, "-gcflags=-dwarflocationlists=true")
-	cmd.Env = append(os.Environ(), `CGO_CFLAGS=-O0 -g3`)
-
-	authdPam := filepath.Join(t.TempDir(), "authd-pam")
-	t.Logf("Compiling Exec client at %s", authdPam)
-	t.Log(strings.Join(cmd.Args, " "))
-
-	cmd.Args = append(cmd.Args, "-o", authdPam)
-	out, err := cmd.CombinedOutput()
-	require.NoError(t, err, "Setup: could not compile PAM client: %s", out)
-
-	return authdPam
-}
-
 func TestMockgpasswd(t *testing.T) {
 	localgroupstestutils.Mockgpasswd(t)
-}
-
-// prependBinToPath returns the value of the GOPATH defined in go env prepended to PATH.
-func prependBinToPath(t *testing.T) string {
-	t.Helper()
-
-	cmd := exec.Command("go", "env", "GOPATH")
-	out, err := cmd.CombinedOutput()
-	require.NoError(t, err, "Could not get GOPATH: %v: %s", err, out)
-
-	env := os.Getenv("PATH")
-	return "PATH=" + strings.Join([]string{filepath.Join(strings.TrimSpace(string(out)), "bin"), env}, ":")
-}
-
-// saveArtifactsForDebug saves the specified artifacts to a temporary directory if the test failed.
-func saveArtifactsForDebug(t *testing.T, artifacts []string) {
-	t.Helper()
-	if !t.Failed() {
-		return
-	}
-
-	// We need to copy the artifacts to another directory, since the test directory will be cleaned up.
-	artifactPath := os.Getenv("AUTHD_TEST_ARTIFACTS_PATH")
-	if artifactPath == "" {
-		artifactPath = filepath.Join(os.TempDir(), "authd-test-artifacts")
-	}
-	tmpDir := filepath.Join(artifactPath, testutils.GoldenPath(t))
-	if err := os.MkdirAll(tmpDir, 0750); err != nil && !os.IsExist(err) {
-		require.NoError(t, err, "Could not create temporary directory for artifacts")
-		return
-	}
-
-	// Copy the artifacts to the temporary directory.
-	for _, artifact := range artifacts {
-		content, err := os.ReadFile(artifact)
-		if err != nil {
-			t.Logf("Could not read artifact %q: %v", artifact, err)
-			continue
-		}
-		if err := os.WriteFile(filepath.Join(tmpDir, filepath.Base(artifact)), content, 0600); err != nil {
-			t.Logf("Could not write artifact %q: %v", artifact, err)
-		}
-	}
 }
